Implement @CalledMethods insertion for AutoValue Builders (#17)
Support for inserting `@This` annotations for AutoValue Builders was added to the returns receiver checker in msridhar/returnsrecv-checker#1.  Building on that, this PR implements `@CalledMethods` insertion for AutoValue Builders.  With this change, the checker is able to check AutoValue Builder usage for correctness with no manual annotations required; see updates to the `Animal.java` test case.