Handle constructors like methods
A PR comment on #207 suggested using a `SimpleAnnotatedTypeScanner` rather than manually replacing only the top-level annotation for non-owning method parameters. That was a mistake.

The original reason that any types were replaced was to avoid false positive warnings when ownership wasn't being transferred. Using a type scanner causes the replacement to happen throughout the whole type - that seems good, but actually is a serious problem because generics are invariant: `@MustCallUnknown List<@MustCallUnknown String>` is NOT a supertype of `@MustCall({}) List<@MustCall({}) String>`, which generates a different spurious false positive.

This PR fixes the implementation and adds some test cases based on FPs in Zookeeper that this problem caused to prevent its recurrence.